Ajustando diretórios, corrigindo fontes e o problema de carregamento do css
