+++ src/app/layout.js
@@ -1,12 +1,4 @@
-import { Inter, Alata } from "next/font/google";
-import "./globals.css";
-
-const inter = Inter({ subsets: ["latin"] });
-const alata = Alata({
-  style: "normal",
-  weight: ["400"],
-  subsets: ["latin"],
-});
+import "../styles/globals.css";
 
 export const metadata = {
   title: "Angelina Cruz | Nutrição & Vida",
@@ -17,9 +9,7 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={`${inter.className} ${alata.className}`}>
-        {children}
-      </body>
+      <body>{children}</body>
     </html>
   );
 }
