+++ src/app/page.js
@@ -1,9 +1,9 @@
 import { HeroSection } from "@/components/HeroSection";
 import { Header } from "@/components/Header";
 import { CaptivationSection } from "@/components/CaptivationSection";
-
-import styles from "./page.module.css";
 import { BenefitsRange } from "@/components/BenefitsRange";
+
+import styles from "../styles/pages/page.module.css";
 
 export default function Home() {
   return (
