+++ src/components/CaptivationSection/CaptivationCard/style.js
@@ -1,3 +1,4 @@
+import React from "react";
 import { styled } from "styled-components";
 
 export const Container = styled.div`
