+++ src/components/CaptivationSection/style.js
@@ -1,3 +1,5 @@
+import React from "react";
+
 import { styled } from "styled-components";
 
 export const Container = styled.section`
