+++ src/components/HeroSection/index.jsx
@@ -1,17 +1,19 @@
 "use client";
-import { Container } from "./style";
+import { inter, alata } from "@/fonts";
+
+import { Container } from "./style.js";
 
 export function HeroSection() {
   return (
     <Container>
       <div>
-        <h2>Alimente seu corpo, nutra sua vida</h2>
-        <p className="title">
+        <h2 style={inter.style}>Alimente seu corpo, nutra sua vida</h2>
+        <p className="title" style={alata.style}>
           DESCUBRA O PODER DE UMA <br /> ALIMENTAÇÃO EQUILIBRADA
         </p>
       </div>
       <div className="containerBrand">
-        <div className="brand">
+        <div className="brand" style={alata.style}>
           <p className="name">Angelina Cruz</p>
           <p className="slogan">NUTRIÇÃO & VIDA</p>
         </div>
