+++ src/styles/globals.css
@@ -1,0 +1,26 @@
+:root {
+  --max-width: 1100px;
+  --border-radius: 12px;
+}
+
+* {
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+}
+
+html,
+body {
+  max-width: 100vw;
+  overflow-x: hidden;
+}
+
+body {
+  color: rgb(var(--foreground-rgb));
+  background-color: pink;
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+}
